Diomira dear (#4)
* Readme for Notebooks directory and DIOMIRA NB

This commit includes several important additions to the Notebooks
respository:
1) A Readme file explaining the mission and scope of the Notebooks.
2) A configuration script to be run before running the Nbs (manage.sh)
3) A Nb describing the city of Diomira (in $NBDIR/invisible_cities/cities/diomira.ipynb)
4) A config file needed to run the Nb (in
$NBDIR/invisible_cities/config/diomira.conf)
5) Two data files, one needed to run Diomira, one produced by
Diomira (but input to next city, Irene, thus needed)

The files are in $NBDIR/invisible_cities/files

* Adding diomira.nb

The NB had to be deleted due to the nasty conflicts that NBs create
with themselves when modified even by their own owner. New version of
the nb now added.