Diomira new (#13)
* deleting locally cities

* New way of saving Notebooks

This commit returns diomira.ipynb to the repository. However, the
output of the nb is stripped before comitting to github. This has two
advantages: it reduces the size of the NB (which can eventually become
quite large when all the output is saved) and it avoids conflicts
associated with the nb changing each time that the output changes.

This implies, inp particular, that different developers can now
collaborate in a given notebook, since the main source of conflict was
the output.

Furthermore, the python (diomira.py) and the html (diomira.html)
versions are committed to git. The .py allows tracking, while the
.html shows the NB otuput.

This commit adds the files .gitattributes, .gitconfig and
ipynb_drop_out which operate all the necessary magic to produce
automatically .py and .html files (each time the nb is saved), and
strip the .ipynb notebook from the output.

* Disable automatic generation of .html and .py

Automatic generation of .html files would result also in modifies
files when output is modified. html and py files must be generated
only in the prod branch and not in the devel branch